+++ bench/bench.js
@@ -14,6 +14,7 @@ import {
   binZsh,
 } from "../test/_constants.js";
 
+import { noShell } from "../src/options.js";
 import * as unix from "../src/unix.js";
 import * as win from "../src/win.js";
 
@@ -27,88 +28,44 @@ const suite = new Benchmark.Suite("escapeShellArg", {
   },
 });
 
+const escapeShellArgNoShellUnix = unix.getEscapeFunction(noShell);
+suite.add(`unix, no shell, ${sampleArg}`, () => {
+  escapeShellArgNoShellUnix(sampleArg);
+});
+
+const escapeShellArgBash = unix.getEscapeFunction(binBash);
 suite.add(`unix, ${binBash}, ${sampleArg}`, () => {
-  const escapeShellArg = unix.getEscapeFunction(binBash, {
-    interpolation: true,
-  });
-  escapeShellArg(sampleArg);
+  escapeShellArgBash(sampleArg);
 });
 
-const escapeShellArgBashNew = unix.getEscapeFunction(binBash, {
-  interpolation: true,
-});
-suite.add(`unix (new), ${binBash}, ${sampleArg}`, () => {
-  escapeShellArgBashNew(sampleArg);
-});
-
+const escapeShellArgCsh = unix.getEscapeFunction(binCsh);
 suite.add(`unix, ${binCsh}, ${sampleArg}`, () => {
-  const escapeShellArg = unix.getEscapeFunction(binCsh, {
-    interpolation: true,
-  });
-  escapeShellArg(sampleArg);
+  escapeShellArgCsh(sampleArg);
 });
 
-const escapeShellArgCshNew = unix.getEscapeFunction(binCsh, {
-  interpolation: true,
-});
-suite.add(`unix (new), ${binCsh}, ${sampleArg}`, () => {
-  escapeShellArgCshNew(sampleArg);
-});
-
+const escapeShellArgDash = unix.getEscapeFunction(binDash);
 suite.add(`unix, ${binDash}, ${sampleArg}`, () => {
-  const escapeShellArg = unix.getEscapeFunction(binDash, {
-    interpolation: true,
-  });
-  escapeShellArg(sampleArg);
+  escapeShellArgDash(sampleArg);
 });
 
-const escapeShellArgDashNew = unix.getEscapeFunction(binDash, {
-  interpolation: true,
-});
-suite.add(`unix (new), ${binDash}, ${sampleArg}`, () => {
-  escapeShellArgDashNew(sampleArg);
-});
-
+const escapeShellArgZsh = unix.getEscapeFunction(binZsh);
 suite.add(`unix, ${binZsh}, ${sampleArg}`, () => {
-  const escapeShellArg = unix.getEscapeFunction(binZsh, {
-    interpolation: true,
-  });
-  escapeShellArg(sampleArg);
+  escapeShellArgZsh(sampleArg);
 });
 
-const escapeShellArgZshNew = unix.getEscapeFunction(binZsh, {
-  interpolation: true,
-});
-suite.add(`unix (new), ${binZsh}, ${sampleArg}`, () => {
-  escapeShellArgZshNew(sampleArg);
+const escapeShellArgNoShellWin = win.getEscapeFunction(noShell);
+suite.add(`win, no shell, ${sampleArg}`, () => {
+  escapeShellArgNoShellWin(sampleArg);
 });
 
+const escapeShellArgCmd = win.getEscapeFunction(binCmd);
 suite.add(`win, ${binCmd}, ${sampleArg}`, () => {
-  const escapeShellArg = win.getEscapeFunction(binCmd, {
-    interpolation: true,
-  });
-  escapeShellArg(sampleArg);
+  escapeShellArgCmd(sampleArg);
 });
 
-const escapeShellArgCmdNew = win.getEscapeFunction(binCmd, {
-  interpolation: true,
-});
-suite.add(`win (new), ${binCmd}, ${sampleArg}`, () => {
-  escapeShellArgCmdNew(sampleArg);
-});
-
+const escapeShellArgPowerShell = win.getEscapeFunction(binPowerShell);
 suite.add(`win, ${binPowerShell}, ${sampleArg}`, () => {
-  const escapeShellArg = win.getEscapeFunction(binPowerShell, {
-    interpolation: true,
-  });
-  escapeShellArg(sampleArg);
-});
-
-const escapeShellArgPowerShellNew = win.getEscapeFunction(binPowerShell, {
-  interpolation: true,
-});
-suite.add(`win (new), ${binPowerShell}, ${sampleArg}`, () => {
-  escapeShellArgPowerShellNew(sampleArg);
+  escapeShellArgPowerShell(sampleArg);
 });
 
 suite.run();