+++ test/integration/escape-all/zsh.test.js
@@ -1,21 +1,20 @@
 /**
- * @overview Contains integration tests for `shescape.escapeAll` for the Z shell
+ * @overview Contains integration tests for `Shescape#escapeAll` for the Z shell
  * (Zsh).
  * @license MIT
  */
 
-import test from "ava";
+import { common, constants, generate } from "../_.js";
 
-import { constants, generate } from "../_.js";
+import { Shescape } from "shescape";
 
-import { escapeAll } from "shescape";
-
-const runTest = constants.isWindows ? test.skip : test;
+const runTest = common.getTestFn(constants.binZsh);
 
 runTest(`input is escaped for ${constants.binZsh}`, (t) => {
   for (const scenario of generate.escapeExamples(constants.binZsh)) {
     const { expected, input, options } = scenario;
-    const result = escapeAll([input], options);
+    const shescape = new Shescape(options);
+    const result = shescape.escapeAll([input]);
     t.deepEqual(result, [expected]);
   }
 });