+++ test/integration/escape/valid.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for valid use of `shescape.escape`.
+ * @overview Contains integration tests for valid use of `Shescape#escape`.
  * @license MIT
  */
 
@@ -7,13 +7,20 @@ import { testProp } from "@fast-check/ava";
 
 import { arbitrary } from "../_.js";
 
-import { escape } from "shescape";
+import { Shescape } from "shescape";
 
 testProp(
   "return values",
   [arbitrary.shescapeArg(), arbitrary.shescapeOptions()],
   (t, arg, options) => {
-    const result = escape(arg, options);
+    let shescape;
+    try {
+      shescape = new Shescape(options);
+    } catch (_) {
+      return t.pass();
+    }
+
+    const result = shescape.escape(arg);
     t.is(typeof result, "string");
   },
 );