+++ test/integration/quote-all/cmd-exe.test.js
@@ -1,21 +1,20 @@
 /**
- * @overview Contains integration tests for `shescape.quoteAll` for the Windows
+ * @overview Contains integration tests for `Shescape#quoteAll` for the Windows
  * Command Prompt (with extension).
  * @license MIT
  */
 
-import test from "ava";
+import { common, constants, generate } from "../_.js";
 
-import { constants, generate } from "../_.js";
+import { Shescape } from "shescape";
 
-import { quoteAll } from "shescape";
-
-const runTest = constants.isWindows ? test : test.skip;
+const runTest = common.getTestFn(constants.binCmd);
 
 runTest(`input is escaped for ${constants.binCmd}`, (t) => {
   for (const scenario of generate.quoteExamples(constants.binCmd)) {
     const { expected, input, options } = scenario;
-    const result = quoteAll([input], options);
+    const shescape = new Shescape(options);
+    const result = shescape.quoteAll([input]);
     t.deepEqual(result, [expected]);
   }
 });