+++ test/unit/options/parse-options.test.js
@@ -11,7 +11,7 @@ import sinon from "sinon";
 import { arbitrary } from "./_.js";
 
 import { resolveExecutable } from "../../../src/executables.js";
-import { parseOptions } from "../../../src/options.js";
+import { noShell, parseOptions } from "../../../src/options.js";
 
 const arbitraryInput = () =>
   fc
@@ -24,15 +24,18 @@ const arbitraryInput = () =>
 test.beforeEach((t) => {
   const getDefaultShell = sinon.stub();
   const getShellName = sinon.stub();
+  const isShellSupported = sinon.stub();
 
-  t.context.deps = { getDefaultShell, getShellName };
+  isShellSupported.returns(true);
+
+  t.context.deps = { getDefaultShell, getShellName, isShellSupported };
 });
 
 testProp("flag protection not specified", [arbitraryInput()], (t, args) => {
   delete args.options.flagProtection;
 
   const result = parseOptions(args, t.context.deps);
-  t.is(result.flagProtection, false);
+  t.is(result.flagProtection, true);
 });
 
 testProp("flag protection set to true", [arbitraryInput()], (t, args) => {
@@ -49,35 +52,79 @@ testProp("flag protection set to false", [arbitraryInput()], (t, args) => {
   t.is(result.flagProtection, false);
 });
 
-testProp("interpolation not specified", [arbitraryInput()], (t, args) => {
-  delete args.options.interpolation;
+testProp(
+  "shell is false",
+  [arbitraryInput(), fc.constantFrom(false)],
+  (t, args, providedShell) => {
+    args.options.shell = providedShell;
 
-  const result = parseOptions(args, t.context.deps);
-  t.is(result.interpolation, false);
-});
-
-testProp("interpolation set to true", [arbitraryInput()], (t, args) => {
-  args.options.interpolation = true;
-
-  const result = parseOptions(args, t.context.deps);
-  t.is(result.interpolation, true);
-});
-
-testProp("interpolation set to false", [arbitraryInput()], (t, args) => {
-  args.options.interpolation = false;
-
-  const result = parseOptions(args, t.context.deps);
-  t.is(result.interpolation, false);
-});
+    const result = parseOptions(args, t.context.deps);
+    t.is(t.context.deps.getDefaultShell.callCount, 0);
+    t.is(t.context.deps.getShellName.callCount, 0);
+    t.is(result.shellName, noShell);
+  },
+);
 
 testProp(
-  "shell is not specified",
+  "shell is omitted",
+  [arbitraryInput(), fc.string(), fc.string()],
+  (t, args, defaultShell, shellName) => {
+    t.context.deps.getDefaultShell.resetHistory();
+    t.context.deps.getDefaultShell.returns(defaultShell);
+    t.context.deps.getShellName.resetHistory();
+    t.context.deps.getShellName.returns(shellName);
+
+    const env = args.env;
+    delete args.options.shell;
+
+    const result = parseOptions(args, t.context.deps);
+    t.is(t.context.deps.getDefaultShell.callCount, 1);
+    t.true(t.context.deps.getDefaultShell.calledWithExactly({ env }));
+    t.is(t.context.deps.getShellName.callCount, 1);
+    t.true(
+      t.context.deps.getShellName.calledWithExactly(
+        { env, shell: defaultShell },
+        { resolveExecutable },
+      ),
+    );
+    t.is(result.shellName, shellName);
+  },
+);
+
+testProp(
+  "shell is nil",
   [
     arbitraryInput(),
-    fc.constantFrom(undefined, true, false),
+    fc.constantFrom(null, undefined),
     fc.string(),
     fc.string(),
   ],
+  (t, args, providedShell, defaultShell, shellName) => {
+    t.context.deps.getDefaultShell.resetHistory();
+    t.context.deps.getDefaultShell.returns(defaultShell);
+    t.context.deps.getShellName.resetHistory();
+    t.context.deps.getShellName.returns(shellName);
+
+    const env = args.env;
+    args.options.shell = providedShell;
+
+    const result = parseOptions(args, t.context.deps);
+    t.is(t.context.deps.getDefaultShell.callCount, 1);
+    t.true(t.context.deps.getDefaultShell.calledWithExactly({ env }));
+    t.is(t.context.deps.getShellName.callCount, 1);
+    t.true(
+      t.context.deps.getShellName.calledWithExactly(
+        { env, shell: defaultShell },
+        { resolveExecutable },
+      ),
+    );
+    t.is(result.shellName, shellName);
+  },
+);
+
+testProp(
+  "shell is truthy",
+  [arbitraryInput(), fc.constantFrom(true), fc.string(), fc.string()],
   (t, args, providedShell, defaultShell, shellName) => {
     t.context.deps.getDefaultShell.resetHistory();
     t.context.deps.getDefaultShell.returns(defaultShell);
@@ -121,5 +168,35 @@ testProp(
       ),
     );
     t.is(result.shellName, shellName);
+  },
+);
+
+testProp("shell is supported", [arbitraryInput()], (t, args) => {
+  t.context.deps.isShellSupported.resetHistory();
+  t.context.deps.isShellSupported.returns(true);
+
+  t.notThrows(() => parseOptions(args, t.context.deps));
+  t.is(t.context.deps.isShellSupported.callCount, 1);
+});
+
+testProp(
+  "shell is unsupported",
+  [
+    arbitraryInput(),
+    fc.oneof(fc.string(), fc.constantFrom(true, null, undefined)),
+    fc.string(),
+  ],
+  (t, args, providedShell, shellName) => {
+    t.context.deps.isShellSupported.resetHistory();
+    t.context.deps.isShellSupported.returns(false);
+    t.context.deps.getShellName.returns(shellName);
+
+    args.options.shell = providedShell;
+
+    t.throws(() => parseOptions(args, t.context.deps), {
+      instanceOf: Error,
+      message: `Shescape does not support the shell ${shellName}`,
+    });
+    t.is(t.context.deps.isShellSupported.callCount, 1);
   },
 );