+++ test/unit/unix/index.test.js
@@ -12,11 +12,13 @@ import sinon from "sinon";
 
 import { arbitrary, constants } from "./_.js";
 
+import * as unix from "../../../src/unix.js";
 import * as bash from "../../../src/unix/bash.js";
 import * as csh from "../../../src/unix/csh.js";
 import * as dash from "../../../src/unix/dash.js";
-import * as unix from "../../../src/unix.js";
+import * as nosh from "../../../src/unix/no-shell.js";
 import * as zsh from "../../../src/unix/zsh.js";
+import { noShell } from "../../../src/options.js";
 
 const shells = [
   { module: bash, shellName: constants.binBash },
@@ -30,30 +32,34 @@ test("the default shell", (t) => {
   t.is(result, "/bin/sh");
 });
 
+test("escape function for no shell", (t) => {
+  const actual = unix.getEscapeFunction(noShell);
+  const expected = nosh.getEscapeFunction();
+  t.deepEqual(actual, expected);
+});
+
 for (const { module, shellName } of shells) {
   test(`escape function for ${shellName}`, (t) => {
-    let options = { interpolation: false };
-    t.is(
-      unix.getEscapeFunction(shellName, options),
-      module.getEscapeFunction(options),
-    );
-
-    options = { interpolation: true };
-    t.is(
-      unix.getEscapeFunction(shellName, options),
-      module.getEscapeFunction(options),
-    );
+    const actual = unix.getEscapeFunction(shellName);
+    const expected = module.getEscapeFunction();
+    t.deepEqual(actual, expected);
   });
 }
 
 testProp(
   "escape function for unsupported shell",
-  [arbitrary.unsupportedUnixShell(), fc.boolean()],
-  (t, shellName, interpolation) => {
-    const result = unix.getEscapeFunction(shellName, { interpolation });
+  [arbitrary.unsupportedUnixShell()],
+  (t, shellName) => {
+    const result = unix.getEscapeFunction(shellName);
     t.is(result, undefined);
   },
 );
+
+test("quote function for no shell", (t) => {
+  const actual = unix.getQuoteFunction(noShell);
+  const expected = nosh.getQuoteFunction();
+  t.deepEqual(actual, expected);
+});
 
 for (const { module, shellName } of shells) {
   test(`quote function for ${shellName}`, (t) => {
@@ -92,7 +98,7 @@ testProp(
     resolveExecutable.returns(path.join(basePath, shell));
 
     const result = unix.getShellName({ env, shell }, { resolveExecutable });
-    t.is(result, constants.binBash);
+    t.is(result, shell);
   },
 );
 
@@ -117,6 +123,12 @@ testProp(
   },
 );
 
+test("flag protection function for no shell", (t) => {
+  const actual = unix.getFlagProtectionFunction(noShell);
+  const expected = nosh.getFlagProtectionFunction();
+  t.is(actual, expected);
+});
+
 for (const { module, shellName } of shells) {
   test(`flag protection function for ${shellName}`, (t) => {
     const actual = unix.getFlagProtectionFunction(shellName);
@@ -131,5 +143,26 @@ testProp(
   (t, shellName) => {
     const result = unix.getFlagProtectionFunction(shellName);
     t.is(result, undefined);
+  },
+);
+
+test(`is shell supported, no shell`, (t) => {
+  const actual = unix.isShellSupported(noShell);
+  t.true(actual);
+});
+
+for (const { shellName } of shells) {
+  test(`is shell supported, ${shellName}`, (t) => {
+    const actual = unix.isShellSupported(shellName);
+    t.true(actual);
+  });
+}
+
+testProp(
+  "is shell supported for unsupported shell",
+  [arbitrary.unsupportedUnixShell()],
+  (t, shellName) => {
+    const result = unix.isShellSupported(shellName);
+    t.false(result);
   },
 );